requirements-desktop.txt
# Dependencies for JARVIS Desktop build
pyinstaller>=6.0.0
pystray>=0.19.0
Pillow>=10.0.0
keyboard>=0.13.5
